SERVER-45930 Add new recommended dependency on mongodb-database-tools
The server tools are not distributed with MongoDB v4.4 and higher
anymore. From now on, there will be a recommended dependency on the
mongodb-database-tools package so that users who expected the tools to be
installed by default will get them, since that is the likely use case
for most users of this Homebrew tap. Users who do not want the tools
installed can now opt out of them.