feat MongoDB Enterprise 8.0 formula for Homebrew
This commit introduces the Homebrew formula for MongoDB Enterprise 8.0. It includes support for both Intel and ARM architectures, with the necessary URLs and SHA256 checksums for each. The formula also handles conflicts with the mongodb-community package and includes caveats regarding the MongoDB Customer Agreement. Additionally, it sets up the plist for the macOS service and provides a default mongod.conf configuration file. An option to enable test commands is available for users who need it. The post_install step ensures the creation of necessary directories for MongoDB operation.